requirements.txt
accelerate==0.21.0
blinker==1.7.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colour==0.1.5
contourpy==1.2.1
cycler==0.12.1
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-cytoscape==1.0.0
dash-html-components==2.0.0
dash-table==5.0.0
diffusers==0.18.2
filelock==3.13.1
Flask==3.0.3
fonttools==4.51.0
fsspec==2023.10.0
gmpy2==2.1.2
h5py==3.11.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==7.0.1
invisible-watermark==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.2
pillow==10.2.0
pip==23.3.1
plotly==5.21.0
psutil==5.9.0
pycparser==2.21
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
safetensors==0.4.2
scipy==1.13.0
setuptools==68.2.2
six==1.16.0
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.65.0
transformers==4.37.2
triton==2.2.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
Werkzeug==3.0.2
wheel==0.41.2
zipp==3.17.0
